requirements.txt
pytest
selenium
requests
Pillow
